pyproject.toml
[tool.poetry]
name = "readflow"
version = "0.1.0"
description = "Reading and listening in a new vision"
authors = ["r-scheele <abdulrahmanolamilekan88@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["email"], version = "^1.10.2"}
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
aioredis = "2.0.0"
injector = "^0.20.1"
fastapi-limiter = {url = "https://github.com/hardbyte/fastapi-limiter/archive/patch-1.zip"}
fastapi-camelcase = "^1.0.5"
beanie = "^1.11.10"
fastapi-utils = "^0.2.1"
bcrypt = "^4.0.0"
PyJWT = "2.1.0"
Pillow = "^9.2.0"
fastapi-mail = "^1.1.4"
Jinja2 = "^3.1.2"
sentry-sdk = "^1.9.9"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
